+++ src/main/java/org/delegator/engine/LoggedUser.java
@@ -2,8 +2,8 @@ package org.delegator.engine;
 
 public interface LoggedUser {
 
-	public int get_UserEid();
-	public void set_UserEid(int userEid);
-	public int isRegistered(String username, String password);
+	public Long get_UserEid();
+	public void set_UserEid(Long userEid);
+	public Long isRegistered(String username, String password);
 
 }
